Change data type of time to BigDecimal

+++ backend/src/main/java/se/dtime/model/report/AccountReport.java
@@ -1,6 +1,7 @@
 package se.dtime.model.report;
 
 import lombok.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -10,5 +11,5 @@ import lombok.*;
 public class AccountReport {
     private Long accountId;
     private String accountName;
-    private double totalHours;
+    private BigDecimal totalHours;
 }
