+++ backend/src/main/java/se/dtime/model/report/BillableTaskTypeReport.java
@@ -2,6 +2,7 @@ package se.dtime.model.report;
 
 import lombok.*;
 import se.dtime.model.TaskType;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -11,7 +12,7 @@ import se.dtime.model.TaskType;
 public class BillableTaskTypeReport {
     private TaskType taskType;
     private Boolean isBillable;
-    private double totalHours;
+    private BigDecimal totalHours;
     private long taskCount;
     private String description;
 }
