+++ backend/src/main/java/se/dtime/model/report/TaskReport.java
@@ -2,6 +2,7 @@ package se.dtime.model.report;
 
 import lombok.*;
 import se.dtime.model.TaskType;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -15,5 +16,5 @@ public class TaskReport {
     private String taskName;
     private TaskType taskType;
     private Boolean isBillable;
-    private double totalHours;
+    private BigDecimal totalHours;
 }
