+++ backend/src/main/java/se/dtime/model/report/UnclosedUser.java
@@ -2,6 +2,7 @@ package se.dtime.model.report;
 
 import lombok.Builder;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -9,6 +10,6 @@ public class UnclosedUser {
     private Long userId;
     private String fullName;
     private String email;
-    private double totalTime;
+    private BigDecimal totalTime;
     private boolean closed;
 }
